First commit
Does compile but assumes _build/handlerbars exists. This was a quick attempt to get the output we need, now I need to wrap this up in a command line utility.

I'm thinking that may need to add parameters for namespace of the module (if we care, currently it takes the base dir the soruce is in, so app in our case) and
a list of input files and maybe an option to either output a single file like it currently does or separate files.