I think this will be enough to get us started.
See usuage details from `--help`.

Current takes an input directory and outputs a single file defined by the out argument with the contents of the compiled templates.

next phase will probably be to split this out into multiple files, but I think that's going to take a little more thought than this version.