More progress, copied some of the models from first.stab and added a world model.
The routes have been defined, stil a work in a progress. In particular the world route isn't completely done. Running into some interesting issues with using aliases instead of id's for like the worlds and exits. I know our goals it to have friendly urls, but I'm starting to wonder if id's would be better. It would jive a little more with the framework I think.

A great example being an exit. When given a route that looks like /world/menn/room/1/edit/exit/south and you have a route defined like

this.resource('exit', { path: '/:room_id/exit' }, function() {
  this.route('edit', { path: '/:direction' });
});

the edit in this example would would have no context for looking up the exit. You would need to know the parent routes room id, to first look up the room the resolve the exist based on that and I feel like that would make things bleed over a bit.

If there was an id instead you could query for the exit with the that id and retrieve the information without needing the parent. Also on the parent route you could handle the willTransition event I think so when going into room/123/exit/5 if the transition will happen, you could interogate the target route (and I think the models have been resolved at that point, not sure) and ensure that 5 is a valid exit for 123.

Spent the last few hours mulling over this in my head and would love some feedback.